+++ Entities/Concrete/Color.cs
@@ -10,7 +10,7 @@ namespace Entities.Concrete
 {
     public class Color:IEntity
     {
-        public int ID { get; set; }
+        public short ID { get; set; }
         public string Name { get; set; }
     }
 }
